agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [ "openai-agents>=0.2.11", "python-dotenv>=1.1.1",]
[[project.authors]]
name = "Marjan Ahmed"
email = "marjan.ahmed08@yahoo.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
agent = "agent:main"
helpful-assistant = "agent.agent:main"
agent-helpful-assistant = "agent.agent:main"
